requirements.txt
flask
faiss-cpu
numpy
openai
python-dotenv
twilio
razorpay